// generated/types/shop_list_create_request.ts
// 由 api-gen.js 生成
// 請勿手動編輯此檔案

export type Shop_list_create_request = {
    page?: number | null;
    page_size?: number | null;
    city?: string | null;
    township?: string | null;
    price_min?: number | null;
    price_max?: number | null;
    keyword?: string | null
};